Added Some Basic Distance Measuring
In lieu of a ruler, I added some text which shows the user the
distance of the currently being dragged or last dragged game object.
In order to check distance without moving, I added the ctrl key.
When pressed, it will update the distance of the game object
being dragged to the mouse pointer without updating its position.

The text is sorta crappy being overlayed in the middle of the
zoomed out grid, but appropriately in the corner when zoomed in.
Really, a separate scene should be made to overlay the gameboards
so that stats like these are unaffected by the main scene's camera.