Conversion to Flask Server and Project Directory Structure
This is in the middle of moving to an app that can spawn multiple
tabletops which are shareable via a url. In doing this, the flask
new project tutorial was followed and the conversion of existing
assets begun. It is pseudo-functional in this state, but each
"tabletop" is the same as no websocket code has been changed.