Hooked Up syncserver.py for Multiple Tabletops
Following the previous changeset, the websocket server can now handle
different tabletops based on a token. There is no security. But
the UI was set up so it's only easy to connect to rooms created
in the database.

May be able to delete the requirements.txt but I haven't tested out
the newer setuptools thing.