Added Text Character Token
In a quick effort to support a variety of icons, and aid in using this
for some language learning pursuits, I added a "character" shape which
creates a new token using the contents of the Character input field.
Beyond English characters, emojis and Chinese characters are examples of
cool uses.